+++ src/server/scripts/Kalimdor/OnyxiasLair/instance_onyxias_lair.cpp
@@ -256,7 +256,7 @@ public:
             }
         }
 
-        bool CheckAchievementCriteriaMeet(uint32 criteria_id, Player const* /*source*/, Unit const* /*target*/ = NULL, uint32 /*miscvalue1*/ = 0)
+        bool CheckAchievementCriteriaMeet(uint32 criteria_id, Player const* /*source*/, Unit const* /*target*/ = NULL, uint64 /*miscvalue1*/ = 0)
         {
             switch (criteria_id)
             {
